Engine: csci gui messages uses intptr_t insted of long

+++ Engine/gui/mylabel.cpp
@@ -54,7 +54,7 @@ int MyLabel::pressedon(int /*mx*/, int /*my*/)
     return 0;
 }
 
-int MyLabel::processmessage(int /*mcode*/, int /*wParam*/, long /*lParam*/)
+int MyLabel::processmessage(int /*mcode*/, int /*wParam*/, intptr_t /*lParam*/)
 {
     return -1;                  // doesn't support messages
 }
